requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.10.0
certifi==2021.10.8
cloudinary==1.29.0
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap3==21.2
django-bootstrap4==22.1
django-braces==1.15.0
django-model-utils==4.2.0
django-star-ratings==0.9.1
djangorestframework==3.13.1
gunicorn==20.1.0
python-decouple==3.6
pytz==2021.3
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
swapper==1.3.0
urllib3==1.26.9
whitenoise==6.0.0
